move test to test_light file
